getting location service and handle permissions done

+++ lib/features/pharmacy/datasource/firebase_storage_datasource.dart
@@ -1,8 +1,9 @@
 import 'package:firebase_storage/firebase_storage.dart';
 
 class FirebaseStorageDatasource {
-  Future<void> deleteImages(List<String> imagesUrl) async {
+  Future<void> deleteImages(List<String?> imagesUrl) async {
     for (var imageUrl in imagesUrl) {
+      if (imageUrl == null) continue;
       await FirebaseStorage.instance.refFromURL(imageUrl).delete();
     }
   }
